+++ app/components/InfoComponent/index.js
@@ -1,1 +1,4 @@
 // finish import and exports
+import InfoBlock from 'InfoBlock';
+
+export { InfoBlock };
