--- app/components/index.js
@@ -1,1 +1,0 @@
-import * from 'GameBlock.js'
