+++ app/containers/HomePage/index.js
@@ -1,0 +1,1 @@
+import Homepage from './HomePage'
