+++ app/containers/component/InfoBlock.js
